Update: Nearly finished Demo
The demo is pretty much complete but there is still a lot more work to do.

Firstly the mobile display hasn't been touched whatsoever so it's going to be fun going through all of that.

Secondly, I need to look at the JavaScript and find something which will auto transition between slides on the merch section. This will bring much more business and attraction as readers will be prompted to interact and see some of the appealing samples we post in that section. You can't really do without it for the final part so I need to sort it.

Once these are done, i'll have finished the demo and can get started on what's actually going to be needed in the real website. It's taken me about 15 - 20 hours to do this demo, it should probably take about 10 for an experienced programmer, hand writing it at least.