Update: Started Mobile Version
Started working on the  mobile version of the website. Whilst i'm at it i've been trying to use the vh,vw, rem, em so it's easier.

Maybe it would be better to use a grid instead for the mobile version but that's a lot of work.

I want to have the title be directly above the slogan on the mobile version because it looks hella clean. I've also go a few other plans for the proms which i'll link below.

https://cdn.discordapp.com/attachments/825123673429639218/1173251902977355876/image.png?ex=656346fb&is=6550d1fb&hm=4d7df76b1eb9b143f2b09be2e2847ff0b994edbeecd95126f56cb7058b5b2948&