Update: Mobile Version
Fixed the mobile version a bit better, formatted the prom-col and have formatted the next section with the merch. Next steps are to figure out how to centre the slider whilst keeping the flex box in column. I know it's something to do with the position: relative/absolute but I can't figure out how to do one without the other taking over the space.

I'll figure it out another day.