Update: Mobile Version
Nearly finished now. Still have the issue of getting the slider centred but everything else is looking good.

The header (i'll finish the title tomorrow) is fully scalable and so is the footer and most of the website.

i'll also have to do a tablet/landscape version but overall it's shaping up and looking pretty flexible for displays (which is what i was struggling to grasp).

it's 2am, time for bed.

Oh, I also find that drawing out the websites really help me. Whenever I can't figure out how to make a grid or flexbox, I watch a tutorial and map out on paper how it's going to work, then go on MDN/Codepen/W3S and try out some of their stuff to try and fully understand how i can make it work. I definitely feel like i'm starting to understand more now how css and html actually work vs how it makes it look. It's incredibly interesting to learn!!!