Update: :hover's
I've added a lot of interactive hovers to make the page more friendly and interactive.

By implementing drop-shadows and translateY's, it's quite easy to give a nice impression of the website "responding" to you in a humane way. I want to check if there's anything else I should :hover before moving on to making sure the website looks ok with it on different displays.