Update: Animations and SVGs
Not fully completed but i've added a lot, the website itself looks clean and the only thing left that i want to do is to edit the text box background with an svg of coffee stains.

This means i'll have to create and edit the SVG myself and implement it behind the div, presumably using z-index.