Finished: Homepage!
I've finished all the homepage and different display versions.

I've done a mobile version (<800px) and a tablet version (801px - 1100px); a mix of both mobile and desktop for more information with a sleeker design, taking out the parts I don't necessarily need from desktop, like i did for mobile.

It's been a long run on this but now i've finally finished it I can start on the other pages and be less adventurous, focusing more on how I will present the content.