Finished: Coffee Alt Displays
Finished updating the display for tablet and mobile display.

I also managed to fix the bug for the footer image which was causing the background to push up from the bottom and show the body background colour. Turns out the padding for one of the links was the wrong syntax in the tablet version. Since it wasn't working it was defaulting to the desktop padding (way too much) and pushing the whole footer section up in the tablet version.

I also replaced the SVG direct input into the HTML with the file (I created it from the SVG code I had inputted) since it kept pushing the website to the side in mobile for some reason. It's a lot easier to keep a track of in file format rather than directly written into the code.

Everything is now finished for those sections only took me about 4 hours this time. I'm getting faster :)