Fixing SVG
The SVG for the footer cannot be the same height as the SVG on the top.

I've overwritten the problems on there so that it can be seen but now i'm trying to manipulate the SVG and it won't change. I think it's falling back on another section which tells it to use another SVG of the same type, so I may have to change both?? It's a bit confusing right now, it shouldn't work that way.