Hero -NavMenu
Navigation menu has been finished for the phone. The animation works well and there are no issues with the overflow or being able to see the menu once it's hidden.

NOTE: If there are any issues where you can see some of or all of the menu, it's highly likely some other part of the page is interfering with the page width, overflowing or being squished on phone are big things to look out for so that you don't get caught out by that bug again!