Coffee/Merch: Layout
All device versions are finished, the SVG's are honed for each width.

Since it was so awkward to add the SVG's onto the sections but make it look like they're seamlessly stuck to it, I have had to make changes to the points they're attached for each device width.

The :hover animations are finished and I've adjusted certain things to make it more device friendly (especially for an extra small width of device).

Next will be doing the data-tooltip to create a hover animation which lets you see additional information. I think having a link to information is a waste of the users and my time. Instead I would rather display additional information when the user indicates they want it.