Scroll Timeline Observer
Added the Javascript to be able to do not just one animation, but two!

I can do one for going upwards and one for going sideways based on what I think is best for each thing. I can also add delays in CSS and change the animation how I like.

Changing the CSS is the next step, it's all laid out how I'll do it so i just have to finish it off when I have more time.